Remove attribute information from the registry
This is part of work to make the `Registry` and its contents more resilient to DDL while connected.

`CompositeType` is now just a convenient subclass of `Type` with no extra information, specifically no more attribute information.

Attribute information can be determined during decoding. During binary decoding the attribute information is provided; including type. During text decoding we assume all attributes are text and rely on string conversion to produce the correct type at the API level.

When encoding types for `Struct` (e.g. `Connection.createStruct`) we use the provided values to determine the attributes and their types. This means that the user must provide exact types since PostgreSQL won’t do any conversion of the attribute types unless they happen to be text.
