+++ src/main/java/com/impossibl/postgres/types/ArrayType.java
@@ -30,10 +30,7 @@ package com.impossibl.postgres.types;
 
 import com.impossibl.postgres.protocol.FieldFormat;
 import com.impossibl.postgres.system.procs.Procs;
-import com.impossibl.postgres.system.tables.PgAttribute;
 import com.impossibl.postgres.system.tables.PgType;
-
-import java.util.Collection;
 
 /**
  * A database array type.
@@ -87,8 +84,8 @@ public class ArrayType extends Type {
   }
 
   @Override
-  public void load(PgType.Row source, Collection<PgAttribute.Row> attrs, Registry registry) {
-    super.load(source, attrs, registry);
+  public void load(PgType.Row source, Registry registry) {
+    super.load(source, registry);
     this.elementType = registry.loadType(source.getElementTypeId());
   }
