+++ src/main/java/com/impossibl/postgres/types/BaseType.java
@@ -30,10 +30,7 @@ package com.impossibl.postgres.types;
 
 import com.impossibl.postgres.protocol.FieldFormat;
 import com.impossibl.postgres.system.procs.Procs;
-import com.impossibl.postgres.system.tables.PgAttribute;
 import com.impossibl.postgres.system.tables.PgType.Row;
-
-import java.util.Collection;
 
 /**
  * A database primitive type
@@ -55,8 +52,8 @@ public class BaseType extends Type {
   }
 
   @Override
-  public void load(Row source, Collection<PgAttribute.Row> attrs, Registry registry) {
-    super.load(source, attrs, registry);
+  public void load(Row source, Registry registry) {
+    super.load(source, registry);
   }
 
 }
