+++ src/main/java/com/impossibl/postgres/types/CompositeType.java
@@ -28,17 +28,7 @@
  */
 package com.impossibl.postgres.types;
 
-import com.impossibl.postgres.protocol.FieldFormat;
-import com.impossibl.postgres.system.tables.PgAttribute;
 import com.impossibl.postgres.system.tables.PgType;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-
-import static java.util.Collections.emptyMap;
 
 /**
  * A database composite type.
@@ -48,152 +38,9 @@ import static java.util.Collections.emptyMap;
  */
 public class CompositeType extends Type {
 
-  /**
-   *  An attribute of the composite type.
-   */
-  public static class Attribute {
-
-    private int number;
-    private String name;
-    private Type type;
-    private boolean nullable;
-    private boolean autoIncrement;
-    private boolean hasDefault;
-    private Map<String, Object> typeModifiers;
-
-    Attribute(int number, String name, Type type,
-              boolean nullable, boolean autoIncrement,
-              boolean hasDefault, Map<String, Object> typeModifiers) {
-      this.number = number;
-      this.name = name;
-      this.type = type;
-      this.nullable = nullable;
-      this.autoIncrement = autoIncrement;
-      this.hasDefault = hasDefault;
-      this.typeModifiers = typeModifiers;
-    }
-
-    public int getNumber() {
-      return number;
-    }
-
-    public String getName() {
-      return name;
-    }
-
-    public Type getType() {
-      return type;
-    }
-
-    public boolean isNullable() {
-      return nullable;
-    }
-
-    public boolean isAutoIncrement() {
-      return autoIncrement;
-    }
-
-    public boolean isHasDefault() {
-      return hasDefault;
-    }
-
-    public Map<String, Object> getTypeModifiers() {
-      return typeModifiers;
-    }
-
-    @Override
-    public String toString() {
-      return name + " : " + type;
-    }
-
-  }
-
-  private List<Attribute> attributes;
-
-  public Attribute getAttribute(int number) {
-
-    //First try the obvious
-    if (number > 0 && number <= attributes.size() && attributes.get(number - 1).number == number) {
-      return attributes.get(number - 1);
-    }
-
-    //Now search all
-    for (Attribute attr : attributes) {
-      if (attr.number == number) {
-        return attr;
-      }
-    }
-
-    return null;
-  }
-
-  public List<Attribute> getAttributes() {
-    return attributes;
-  }
-
-  public Type[] getAttributesTypes() {
-    Type[] attributeTypes = new Type[attributes.size()];
-    for (int c = 0; c < attributes.size(); ++c) {
-      attributeTypes[c] = attributes.get(c).type;
-    }
-    return attributeTypes;
-  }
-
   @Override
-  public boolean isParameterFormatSupported(FieldFormat format) {
-
-    boolean allSupported = super.isParameterFormatSupported(format);
-    for (Attribute attr : attributes) {
-      allSupported &= attr.type.isParameterFormatSupported(format);
-    }
-
-    return allSupported;
-  }
-
-  @Override
-  public boolean isResultFormatSupported(FieldFormat format) {
-
-    boolean allSupported = super.isResultFormatSupported(format);
-    for (Attribute attr : attributes) {
-      allSupported &= attr.type.isResultFormatSupported(format);
-    }
-
-    return allSupported;
-  }
-
-  @Override
-  public void load(PgType.Row pgType, Collection<PgAttribute.Row> pgAttrs, Registry registry) {
-    super.load(pgType, pgAttrs, registry);
-
-    if (pgAttrs == null) {
-
-      attributes = Collections.emptyList();
-    }
-    else {
-
-      attributes = new ArrayList<>(pgAttrs.size());
-
-      for (PgAttribute.Row pgAttr : pgAttrs) {
-        Type type = registry.loadType(pgAttr.getTypeId());
-        Attribute attr = new Attribute(pgAttr.getNumber(),
-                                       pgAttr.getName(),
-                                       type,
-                                       pgAttr.isNullable(),
-                                       pgAttr.isAutoIncrement(),
-                                       pgAttr.isHasDefault(),
-                                       type != null ? type.getModifierParser().parse(pgAttr.getTypeModifier()) : emptyMap());
-
-        attributes.add(attr);
-      }
-
-      attributes.sort((o1, o2) -> {
-        int o1n = o1.number < 0 ? o1.number + Integer.MAX_VALUE : o1.number;
-        int o2n = o2.number < 0 ? o2.number + Integer.MAX_VALUE : o2.number;
-        return o1n - o2n;
-      });
-
-    }
-
+  public void load(PgType.Row pgType, Registry registry) {
+    super.load(pgType, registry);
   }
 
 }
