+++ src/main/java/com/impossibl/postgres/types/PsuedoType.java
@@ -28,10 +28,7 @@
  */
 package com.impossibl.postgres.types;
 
-import com.impossibl.postgres.system.tables.PgAttribute;
 import com.impossibl.postgres.system.tables.PgType.Row;
-
-import java.util.Collection;
 
 /**
  * A database psuedo type.
@@ -42,8 +39,8 @@ import java.util.Collection;
 public class PsuedoType extends Type {
 
   @Override
-  public void load(Row source, Collection<PgAttribute.Row> attrs, Registry registry) {
-    super.load(source, attrs, registry);
+  public void load(Row source, Registry registry) {
+    super.load(source, registry);
   }
 
 }
