+++ src/main/java/com/impossibl/postgres/types/RangeType.java
@@ -29,10 +29,7 @@
 package com.impossibl.postgres.types;
 
 import com.impossibl.postgres.protocol.FieldFormat;
-import com.impossibl.postgres.system.tables.PgAttribute;
 import com.impossibl.postgres.system.tables.PgType;
-
-import java.util.Collection;
 
 /**
  * A database range type.
@@ -64,8 +61,8 @@ public class RangeType extends Type {
   }
 
   @Override
-  public void load(PgType.Row source, Collection<PgAttribute.Row> attrs, Registry registry) {
-    super.load(source, attrs, registry);
+  public void load(PgType.Row source, Registry registry) {
+    super.load(source, registry);
     base = registry.loadType(source.getRangeBaseTypeId());
   }
